+++ spec/dummy/app/models/test_customer.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+# A generic polymorphic-owner target for Ledger::Account specs
+class TestCustomer < ApplicationRecord
+end